requirements.txt
Flask==2.1.1
Pillow==9.1.0
Werkzeug==2.0.3
keras